+++ cmd/traffic/cmd/manager/internal/cluster/info.go
@@ -56,7 +56,7 @@ type info struct {
 
 func NewInfo(ctx context.Context) Info {
 	env := managerutil.GetEnv(ctx)
-	managedNamespaces := env.GetManagedNamespaces()
+	managedNamespaces := env.ManagedNamespaces
 	namespaced := len(managedNamespaces) > 0
 	oi := info{}
 	ki := k8sapi.GetK8sInterface(ctx)
@@ -89,47 +89,6 @@ func NewInfo(ctx context.Context) Info {
 		oi.clusterID = license.ClusterIDZero
 		dlog.Warnf(ctx, "unable to get namespace \"default\", will use default clusterID: %s: %v",
 			oi.clusterID, err)
-	}
-
-	// places to look for the cluster's DNS service
-	dnsServices := []metav1.ObjectMeta{
-		{
-			Name:      "kube-dns",
-			Namespace: "kube-system",
-		},
-		{
-			Name:      "coredns",
-			Namespace: "kube-system",
-		},
-		{
-			Name:      "dns-default",
-			Namespace: "openshift-dns",
-		},
-		{
-			Name:      env.DNSServiceName,
-			Namespace: env.DNSServiceNamespace,
-		},
-	}
-
-	var kubeDnsIp net.IP
-	for _, svc := range dnsServices {
-		if ips, err := net.DefaultResolver.LookupIP(ctx, "ip4", svc.Name+"."+svc.Namespace); err == nil && len(ips) > 0 {
-			dlog.Infof(ctx, "Using DNS IP from %s.%s", svc.Name, svc.Namespace)
-			kubeDnsIp = ips[0]
-			break
-		}
-	}
-
-	if kubeDnsIp == nil && env.DNSServiceIP != "" {
-		dlog.Infof(ctx, "Unable to determine DNS IP, using user supplied IP %s", env.DNSServiceIP)
-		kubeDnsIp = net.ParseIP(env.DNSServiceIP)
-		if kubeDnsIp == nil {
-			dlog.Warn(ctx, "The user supplied IP is not a valid IP address")
-		}
-	}
-
-	if kubeDnsIp == nil {
-		dlog.Warn(ctx, "Could not determine DNS ClusterIP")
 	}
 
 	apiSvc := "kubernetes.default.svc"
@@ -175,34 +134,35 @@ func NewInfo(ctx context.Context) Info {
 		}
 	}
 
-	if oi.ServiceSubnet == nil && kubeDnsIp != nil {
+	if err != nil {
+		dlog.Warn(ctx, err)
+	}
+	if oi.ServiceSubnet == nil {
 		// Using a "kubectl cluster-info dump" or scanning all services generates a lot of unwanted traffic
 		// and would quite possibly also require elevated permissions, so instead, we derive the service subnet
-		// from the kubeDNS IP. This is cheating but a cluster may only have one service subnet and the mask is
-		// unlikely to cover less than half the bits.
-		dlog.Infof(ctx, "Deriving serviceSubnet from %s (the IP of kube-dns.kube-system)", kubeDnsIp)
-		bits := len(kubeDnsIp) * 8
-		ones := bits / 2
-		mask := net.CIDRMask(ones, bits) // will yield a 16 bit mask on IPv4 and 64 bit mask on IPv6.
-		oi.ServiceSubnet = &rpc.IPNet{Ip: kubeDnsIp.Mask(mask), Mask: int32(ones)}
+		// from the traffic-manager service IP. This is cheating but a cluster may only have one service subnet
+		// and the mask is unlikely to cover less than half the bits.
+		ips, err := net.DefaultResolver.LookupIP(ctx, "ip4", "traffic-manager")
+		if err != nil || len(ips) == 0 {
+			dlog.Warn(ctx, "traffic manager is not able to resolve the IP of its own service")
+		} else {
+			ip := ips[0]
+			dlog.Infof(ctx, "Deriving serviceSubnet from %s (the IP of traffic-manager.%s)", ip, env.ManagerNamespace)
+			bits := len(ip) * 8
+			ones := bits / 2
+			mask := net.CIDRMask(ones, bits) // will yield a 16 bit mask on IPv4 and 64 bit mask on IPv6.
+			oi.ServiceSubnet = &rpc.IPNet{Ip: ip.Mask(mask), Mask: int32(ones)}
+		}
 	}
 
 	podCIDRStrategy := env.PodCIDRStrategy
 	dlog.Infof(ctx, "Using podCIDRStrategy: %s", podCIDRStrategy)
 
-	oi.ManagerPodIp = iputil.Parse(env.PodIP)
-	if oi.ManagerPodIp == nil {
-		dlog.Warnf(ctx, "Unable to get manager pod ip; env var says %s", env.PodIP)
-	}
+	oi.ManagerPodIp = env.PodIP
+	oi.ManagerPodPort = int32(env.ServerPort)
 
-	alsoProxy, err := env.GetAlsoProxySubnets()
-	if err != nil {
-		dlog.Errorf(ctx, "AlsoProxySubnets not parsed: %v", err)
-	}
-	neverProxy, err := env.GetNeverProxySubnets()
-	if err != nil {
-		dlog.Errorf(ctx, "NeverProxySubnets not parsed: %v", err)
-	}
+	alsoProxy := env.ClientRoutingAlsoProxySubnets
+	neverProxy := env.ClientRoutingNeverProxySubnets
 	dlog.Infof(ctx, "Using AlsoProxy: %v", alsoProxy)
 	dlog.Infof(ctx, "Using NeverProxy: %v", neverProxy)
 
@@ -220,7 +180,7 @@ func NewInfo(ctx context.Context) Info {
 	oi.Dns = &rpc.DNS{
 		IncludeSuffixes: env.ClientDnsIncludeSuffixes,
 		ExcludeSuffixes: env.ClientDnsExcludeSuffixes,
-		KubeIp:          kubeDnsIp,
+		KubeIp:          env.PodIP,
 		ClusterDomain:   clusterDomain,
 	}
 
@@ -317,30 +277,13 @@ func (oi *info) watchPodSubnets(ctx context.Context, namespaces []string) bool {
 }
 
 func (oi *info) setSubnetsFromEnv(ctx context.Context) bool {
-	pcEnv := managerutil.GetEnv(ctx).PodCIDRs
-	cidrStrs := strings.Split(pcEnv, " ")
-	allOK := len(cidrStrs) > 0
-	subnets := make(subnet.Set, len(cidrStrs))
-	if allOK {
-		for _, s := range cidrStrs {
-			_, cidr, err := net.ParseCIDR(s)
-			if err != nil {
-				dlog.Errorf(ctx, "unable to parse CIDR %q from environment variable POD_CIDRS: %v", s, err)
-				allOK = false
-				break
-			}
-			subnets.Add(cidr)
-		}
-	}
-	if allOK {
+	subnets := managerutil.GetEnv(ctx).PodCIDRs
+	if len(subnets) > 0 {
+		oi.PodSubnets = subnetsToRPC(subnets)
 		dlog.Infof(ctx, "Using subnets from POD_CIDRS environment variable")
-		oi.PodSubnets = toRPCSubnets(subnets)
-
-		oi.ciSubs.notify(ctx, oi.clusterInfo())
-	} else {
-		dlog.Errorf(ctx, "unable to parse subnets from POD_CIDRS value %q", pcEnv)
+		return true
 	}
-	return allOK
+	return false
 }
 
 // Watch will start by sending an initial snapshot of the ClusterInfo on the given stream
@@ -355,13 +298,14 @@ func (oi *info) GetClusterID() string {
 
 func (oi *info) clusterInfo() *rpc.ClusterInfo {
 	ci := &rpc.ClusterInfo{
-		ServiceSubnet: oi.ServiceSubnet,
-		PodSubnets:    make([]*rpc.IPNet, len(oi.PodSubnets)),
-		ManagerPodIp:  oi.ManagerPodIp,
-		Routing:       oi.Routing,
-		Dns:           oi.Dns,
-		KubeDnsIp:     oi.Dns.KubeIp,
-		ClusterDomain: oi.Dns.ClusterDomain,
+		ServiceSubnet:  oi.ServiceSubnet,
+		PodSubnets:     make([]*rpc.IPNet, len(oi.PodSubnets)),
+		ManagerPodIp:   oi.ManagerPodIp,
+		ManagerPodPort: oi.ManagerPodPort,
+		Routing:        oi.Routing,
+		Dns:            oi.Dns,
+		KubeDnsIp:      oi.Dns.KubeIp,
+		ClusterDomain:  oi.Dns.ClusterDomain,
 	}
 	copy(ci.PodSubnets, oi.PodSubnets)
 	return ci
@@ -369,13 +313,16 @@ func (oi *info) clusterInfo() *rpc.ClusterInfo {
 
 func (oi *info) watchSubnets(ctx context.Context, retriever subnetRetriever) {
 	retriever.changeNotifier(ctx, func(subnets subnet.Set) {
-		oi.PodSubnets = toRPCSubnets(subnets)
+		oi.PodSubnets = subnetSetToRPC(subnets)
 		oi.ciSubs.notify(ctx, oi.clusterInfo())
 	})
 }
 
-func toRPCSubnets(cidrMap subnet.Set) []*rpc.IPNet {
-	subnets := cidrMap.AppendSortedTo(nil)
+func subnetSetToRPC(cidrMap subnet.Set) []*rpc.IPNet {
+	return subnetsToRPC(cidrMap.AppendSortedTo(nil))
+}
+
+func subnetsToRPC(subnets []*net.IPNet) []*rpc.IPNet {
 	rpcSubnets := make([]*rpc.IPNet, len(subnets))
 	for i, s := range subnets {
 		rpcSubnets[i] = iputil.IPNetToRPC(s)