+++ cmd/traffic/cmd/manager/internal/mutator/agent_injector_test.go
@@ -43,15 +43,12 @@ func stringP(s string) *string {
 
 func TestTrafficAgentConfigGenerator(t *testing.T) {
 	env := &managerutil.Env{
-		User:        "",
-		ServerHost:  "tel-example",
-		ServerPort:  "80",
-		SystemAHost: "",
-		SystemAPort: "",
+		ServerHost: "tel-example",
+		ServerPort: 8081,
 
 		ManagerNamespace: "default",
 		AgentRegistry:    "docker.io/datawire",
-		AgentImage:       "tel2:2.6.0",
+		AgentImage:       "tel2:2.8.0",
 		AgentPort:        9900,
 	}
 	ctx := dlog.NewTestContext(t, false)
@@ -460,7 +457,7 @@ func TestTrafficAgentConfigGenerator(t *testing.T) {
 					Containers: []core.Container{
 						{
 							Ports: []core.ContainerPort{
-								{Name: "http", ContainerPort: env.AgentPort},
+								{Name: "http", ContainerPort: int32(env.AgentPort)},
 							}},
 					},
 				},
@@ -741,11 +738,8 @@ func TestTrafficAgentConfigGenerator(t *testing.T) {
 
 func TestTrafficAgentInjector(t *testing.T) {
 	env := &managerutil.Env{
-		User:        "",
-		ServerHost:  "tel-example",
-		ServerPort:  "80",
-		SystemAHost: "",
-		SystemAPort: "",
+		ServerHost: "tel-example",
+		ServerPort: 8081,
 
 		ManagerNamespace:  "default",
 		AgentRegistry:     "docker.io/datawire",
@@ -1639,7 +1633,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 				ae := reflect.ValueOf(test.envAdditions).Elem()
 				for i := ae.NumField() - 1; i >= 0; i-- {
 					ef := ae.Field(i)
-					if (ef.Kind() == reflect.String || ef.Kind() == reflect.Int32) && !ef.IsZero() {
+					if (ef.Kind() == reflect.String || ef.Kind() == reflect.Uint16) && !ef.IsZero() {
 						ne.Field(i).Set(ef)
 					}
 				}