+++ cmd/traffic/cmd/manager/managerutil/envconfig_test.go
@@ -3,47 +3,54 @@ package managerutil_test
 import (
 	"context"
 	"fmt"
-	"os"
+	"net"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/api/resource"
 
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/managerutil"
+	"github.com/telepresenceio/telepresence/v2/pkg/agentconfig"
+	"github.com/telepresenceio/telepresence/v2/pkg/k8sapi"
 	"github.com/telepresenceio/telepresence/v2/pkg/version"
 )
 
 func TestEnvconfig(t *testing.T) {
-	origEnv := os.Environ()
-	defer func() {
-		os.Clearenv()
-		for _, kv := range origEnv {
-			eq := strings.Index(kv, "=")
-			if eq < 0 {
-				continue
-			}
-			k := kv[:eq]
-			v := kv[eq+1:]
-			os.Setenv(k, v)
-		}
-	}()
+	// Default environment, providing what's necessary for the traffic-manager
+	env := map[string]string{
+		"AGENT_APP_PROTO_STRATEGY":    k8sapi.Http2Probe.String(),
+		"AGENT_ENVOY_ADMIN_PORT":      "19000",
+		"AGENT_ENVOY_SERVER_PORT":     "18000",
+		"AGENT_INJECT_POLICY":         agentconfig.OnDemand.String(),
+		"AGENT_PORT":                  "9900",
+		"AGENT_REGISTRY":              "docker.io/datawire",
+		"CLIENT_CONNECTION_TTL":       (24 * time.Hour).String(),
+		"CLIENT_DNS_EXCLUDE_SUFFIXES": ".com .io .net .org .ru",
+		"GRPC_MAX_RECEIVE_SIZE":       "4Mi",
+		"LOG_LEVEL":                   "info",
+		"POD_IP":                      "203.0.113.18",
+		"POD_CIDR_STRATEGY":           "auto",
+		"SERVER_PORT":                 "8081",
+	}
 
 	defaults := managerutil.Env{
-		ServerPort:               "8081",
-		PrometheusPort:           "0",
-		SystemAHost:              "app.getambassador.io",
-		SystemAPort:              "443",
-		AgentRegistry:            "docker.io/datawire",
+		AgentAppProtocolStrategy: k8sapi.Http2Probe,
+		AgentEnvoyAdminPort:      19000,
+		AgentEnvoyLogLevel:       "info",
+		AgentEnvoyServerPort:     18000,
+		AgentLogLevel:            "info",
 		AgentPort:                9900,
-		MaxReceiveSize:           resource.MustParse("4Mi"),
-		PodCIDRStrategy:          "auto",
-		DNSServiceName:           "coredns",
-		DNSServiceNamespace:      "kube-system",
-		LogLevel:                 "info",
+		AgentRegistry:            "docker.io/datawire",
 		ClientConnectionTTL:      24 * time.Hour,
 		ClientDnsExcludeSuffixes: []string{".com", ".io", ".net", ".org", ".ru"},
+		LogLevel:                 "info",
+		MaxReceiveSize:           resource.MustParse("4Mi"),
+		PodCIDRStrategy:          "auto",
+		PodIP:                    net.IP{203, 0, 113, 18},
+		ServerPort:               8081,
 	}
 
 	testcases := map[string]struct {
@@ -62,29 +69,42 @@ func TestEnvconfig(t *testing.T) {
 				e.SystemAHost = "app.getambassador.io"
 			},
 		},
+		"complex": {
+			Input: map[string]string{
+				"CLIENT_ROUTING_NEVER_PROXY_SUBNETS": "10.20.30.0/24 10.20.40.0/24",
+			},
+			Output: func(e *managerutil.Env) {
+				_, a, _ := net.ParseCIDR("10.20.30.0/24")
+				_, b, _ := net.ParseCIDR("10.20.40.0/24")
+				e.ClientRoutingNeverProxySubnets = []*net.IPNet{a, b}
+			},
+		},
 	}
 
 	for tcName, tc := range testcases {
 		tc := tc // Capture loop variable...
-		// Because we don't run the subtests in parallel, capturing the loop variable
-		// doesn't really matter, but scopelint complains.
-
 		t.Run(tcName, func(t *testing.T) {
-			assert := assert.New(t)
-
-			os.Clearenv()
-			for k, v := range tc.Input {
-				os.Setenv(k, v)
+			t.Parallel()
+			lookup := func(key string) (string, bool) {
+				val, ok := tc.Input[key]
+				if !ok {
+					val, ok = env[key]
+				}
+				return val, ok
 			}
 
 			expected := defaults
 			tc.Output(&expected)
 
-			ctx, err := managerutil.LoadEnv(context.Background())
-			assert.Nil(err)
+			ctx, err := managerutil.LoadEnv(context.Background(), lookup)
+			require.NoError(t, err)
 			actual := managerutil.GetEnv(ctx)
-			assert.Equal(&expected, actual)
-			assert.Equal(fmt.Sprintf("%s/tel2:%s", actual.AgentRegistry, strings.TrimPrefix(version.Version, "v")),
+			assert.Equal(t, &expected, actual)
+			assert.Equal(t,
+				fmt.Sprintf(
+					"%s/tel2:%s",
+					actual.AgentRegistry,
+					strings.TrimPrefix(version.Version, "v")),
 				actual.QualifiedAgentImage())
 		})
 	}