+++ pkg/a8rcloud/systema.go
@@ -3,6 +3,7 @@ package a8rcloud
 import (
 	"context"
 	"crypto/tls"
+	"errors"
 	"net"
 	"sync"
 
@@ -61,11 +62,11 @@ func WithSystemAPool[T Closeable](ctx context.Context, poolName string, provider
 	return context.WithValue(ctx, key, &systemAPool[T]{Provider: provider, Name: poolName, parentCtx: ctx})
 }
 
-func GetSystemAPool[T Closeable](ctx context.Context, poolName string) SystemAPool[T] {
+func GetSystemAPool[T Closeable](ctx context.Context, poolName string) (SystemAPool[T], error) {
 	if p, ok := ctx.Value(systemaPoolKey(poolName)).(*systemAPool[T]); ok {
-		return p
+		return p, nil
 	}
-	return nil
+	return nil, errors.New("access to Ambassador Cloud is not configured")
 }
 
 func GetSystemAPoolProvider[T Closeable](ctx context.Context, poolName string) ClientProvider[T] {