+++ pkg/agentconfig/sidecar.go
@@ -125,13 +125,22 @@ type Sidecar struct {
 	ManagerHost string `json:"managerHost,omitempty"`
 
 	// The port used when connecting to the traffic manager
-	ManagerPort int32 `json:"managerPort,omitempty"`
+	ManagerPort uint16 `json:"managerPort,omitempty"`
 
 	// The port used by the agents restFUL API server
 	APIPort uint16 `json:"apiPort,omitempty"`
 
 	// The port used by the agent's GRPC tracing server
 	TracingPort uint16 `json:"tracingPort,omitempty"`
+
+	// LogLevel used by the envoy instance
+	EnvoyLogLevel string
+
+	// The port used by the Envoy server
+	EnvoyServerPort uint16
+
+	// The port used for Envoy administration
+	EnvoyAdminPort uint16
 
 	// Resources for the sidecar
 	Resources *core.ResourceRequirements `json:"resources,omitempty"`