+++ CHANGELOG.md
@@ -11,8 +11,21 @@
 - Feature: A `dns` struct container the fields `includeSuffixes` and `excludeSuffixes` was added to the Helm 
   chart `client` struct, making those values configurable per cluster.
 
+- Feature: The API port used by the traffic-manager is now configurable using the Helm chart value `apiPort`.
+  The default port is 8081.
+
 - Change: The Helm chart `dnsConfig` was deprecated but retained for backward compatibility. The fields
   `alwaysProxySubnets` and `neverProxySubnets` can now be found under `routing` in the `client` struct.
+
+- Change: The Helm chart `agentInjector.agentImage` was moved to `agent.image`. The old value is deprecated but
+  retained for backward compatibility.
+
+- Change: The Helm chart `agentInjector.appProtocolStrategy` was moved to `agent.appProtocolStrategy`. The old 
+  value is deprecated but retained for backward compatibility.
+
+- Change: The Helm chart `dnsServiceName`, `dnsServiceNamespace`, and `dnsServiceIP` has been removed, because
+  they are no longer needed. The TUN-device will use the traffic-manager pod-IP on platforms where it needs to
+  dedicate an IP for its local resolver.
 
 - Bugfix: Environment variable interpolation now works for all definitions that are copied from pod containers
   into the injected traffic-agent container.